# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1",
    "title": "MongoDB Connector API",
    "description": "A fully-managed API to read, write, and aggregate data in MongoDB.\n"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/action/findOne": {
      "post": {
        "operationId": "findOne",
        "summary": "Find One Document",
        "description": "Find a single document that matches a query.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/FindOneRequestBody"
                  }
                ],
                "example": {
                  "dataSource": "mongodb",
                  "database": "todo",
                  "collection": "tasks",
                  "filter": {
                    "text": "Do the dishes"
                  },
                  "projection": {
                    "status": 1,
                    "text": 1
                  }
                }
              }
            },
            "application/ejson": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/FindOneRequestBody"
                  }
                ],
                "example": {
                  "dataSource": "mongodb",
                  "database": "todo",
                  "collection": "tasks",
                  "filter": {
                    "text": "Do the dishes"
                  },
                  "projection": {
                    "status": 1,
                    "text": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindOneResponseBody"
                },
                "example": {
                  "document": {
                    "_id": "6193504e1be4ab27791c8133",
                    "status": "open",
                    "text": "Do the dishes"
                  }
                }
              },
              "application/ejson": {
                "schema": {
                  "$ref": "#/components/schemas/FindOneResponseBody"
                },
                "example": {
                  "document": {
                    "_id": {
                      "$oid": "6193504e1be4ab27791c8133"
                    },
                    "status": "open",
                    "text": "Do the dishes"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "description": "Unauthorized",
            "$ref": "#/components/responses/UnauthorizedRequestError"
          }
        }
      }
    },
...
    "/action/insertOne": {
      "post": {
        "operationId": "insertOne",
        "summary": "Insert One Document",
        "description": "Insert a single document into a collection.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/InsertOneRequestBody"
                  }
                ],
                "example": {
                  "dataSource": "mongodb",
                  "database": "todo",
                  "collection": "tasks",
                  "document": {
                    "status": "open${{ env.BUNDLED_SPEC_FILEPATH }}",
                    "text": "Do the dishes"
                  }
                }
              }
            },
            "application/ejson": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/InsertOneRequestBody"
                  }
                ],
                "example": {
                  "dataSource": "mongodb",
                  "database": "todo",
                  "collection": "tasks",
                  "document": {
                    "status": "open",
                    "text": "Do the dishes"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inserted",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/InsertOneResponseBody"
                    },
                    {
                      "properties": {
                        "insertedId": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "insertedId": "6193504e1be4ab27791c8133"
                }
              },
              "application/ejson": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/InsertOneResponseBody"
                    },
                    {
                      "properties": {
                        "insertedId": {
                          "type": "object"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "insertedId": {
                    "$oid": "6193504e1be4ab27791c8133"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "description": "Unauthorized",
            "$ref": "#/components/responses/UnauthorizedRequestError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Namespace": {
        "type": "object",
        "required": [
          "dataSource",
          "database",
          "collection"
        ],
        "properties": {
          "dataSource": {
            "type": "string",
            "description": "The name of a linked MongoDB Atlas data source. This is\ncommonly `\"mongodb\"` though it may be different in\nyour App if you chose a different name when you created the\ndata source.\n"
          },
          "database": {
            "type": "string",
            "description": "The name of a database in the specified data source."
          },
          "collection": {
            "type": "string",
            "description": "The name of a collection in the specified database."
          }
        }
      },
      "Filter": {
        "type": "object",
        "properties": {
          "filter": {
            "type": "object",
            "description": "A MongoDB query filter that matches documents. For a list of all query operators that the Data API supports."
          }
        }
      },
      "Projection": {
        "type": "object",
        "properties": {
          "projection": {
            "type": "object",
            "additionalProperties": {
              "type": "number",
              "enum": [
                0,
                1
              ]
            },
            "description": "A MongoDB projection for matched documents returned by the operation."
          }
        }
      },
      "FindOneRequestBody": {
        "title": "FindOneRequestBody",
        "required": [
          "filter"
        ],
        "allOf": [
          {
            "$ref": "#/components/schemas/Namespace"
          },
          {
            "$ref": "#/components/schemas/Filter"
          },
          {
            "$ref": "#/components/schemas/Projection"
          }
        ]
      },
      "FindOneResponseBody": {
        "title": "FindOneResponseBody",
        "type": "object",
        "description": "The result of a findOne operation.",
        "properties": {
          "document": {
            "type": "object",
            "description": "A document that matches the specified filter. If no documents match, this is `null`."
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "A message that describes the error."
          },
          "error_code": {
            "type": "string",
            "description": "The error type."
          },
          "link": {
            "type": "string",
            "description": "A link to a log entry for the failed operation."
          }
        }
      },
      "ErrorNoAuthenticationSpecified": {
        "description": "Indicates that the request did not include any authentication\ncredentials.",
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          }
        ]
      },
      "ErrorMissingAuthenticationParameter": {
        "description": "Indicates that an authentication parameter is missing from the\nrequest body or headers. For example, the request might include\nan `email` header but no corresponding `password` header.",
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          }
        ]
      },
      "ErrorUserNotFound": {
        "description": "Indicates that no user matched the provided authentication\ncredentials.",
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          }
        ]
      },
...
      "InsertOneRequestBody": {
        "title": "InsertOneRequestBody",
        "allOf": [
          {
            "$ref": "#/components/schemas/Namespace"
          },
          {
            "type": "object",
            "required": [
              "document"
            ],
            "properties": {
              "document": {
                "type": "object",
                "description": "A document to insert into the collection."
              }
            }
          }
        ]
      },
      "InsertOneResponseBody": {
        "title": "InsertOneResponseBody",
        "description": "The result of an insertOne operation.",
        "type": "object",
        "properties": {
          "insertedId": {
            "description": "The `_id` value of the inserted document."
          }
        }
      },
...
    },
    "responses": {
      "BadRequestError": {
        "description": "The request was malformed or incomplete.",
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ErrorNoAuthenticationSpecified"
                },
                {
                  "$ref": "#/components/schemas/ErrorMissingAuthenticationParameter"
                }
              ]
            }
          }
        }
      },
      "UnauthorizedRequestError": {
        "description": "The authenticated user does not have permission to access this endpoint.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorUserNotFound"
            }
          }
        }
      }
    }
  }
}